added utility file for report generation, performance metrics integrated to analysis
